requirements.txt
Django==1.11.6
djangorestframework==3.8.2
psycopg2==2.7.3.1
pytz==2017.2
